Updated structure of counter
